// src/gen/validation.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v1
 */
import zod from 'zod';

/**
 * @summary Get all Persons
 */
export const listPersons200ResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "age": zod.number().optional(),
  "address": zod.string().optional(),
  "work": zod.string().optional()
})
export const listPersons200Response = zod.array(listPersons200ResponseItem)


/**
 * @summary Create new Person
 */
export const createPersonBody = zod.object({
  "name": zod.string(),
  "age": zod.number().optional(),
  "address": zod.string().optional(),
  "work": zod.string().optional()
})

export const createPerson400Response = zod.object({
  "message": zod.string().optional(),
  "errors": zod.record(zod.string(), zod.string()).optional()
})


/**
 * @summary Get Person by ID
 */
export const getPersonParams = zod.object({
  "id": zod.coerce.number()
})

export const getPerson200Response = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "age": zod.number().optional(),
  "address": zod.string().optional(),
  "work": zod.string().optional()
})

export const getPerson404Response = zod.object({
  "message": zod.string().optional()
})


/**
 * @summary Remove Person by ID
 */
export const deletePersonParams = zod.object({
  "id": zod.coerce.number()
})


/**
 * @summary Update Person by ID
 */
export const editPersonParams = zod.object({
  "id": zod.coerce.number()
})

export const editPersonBody = zod.object({
  "name": zod.string(),
  "age": zod.number().optional(),
  "address": zod.string().optional(),
  "work": zod.string().optional()
})

export const editPerson200Response = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "age": zod.number().optional(),
  "address": zod.string().optional(),
  "work": zod.string().optional()
})

export const editPerson400Response = zod.object({
  "message": zod.string().optional(),
  "errors": zod.record(zod.string(), zod.string()).optional()
})

export const editPerson404Response = zod.object({
  "message": zod.string().optional()
})